+++ src-input/duk_api_call.c
@@ -1,16 +1,88 @@
 /*
  *  Calls.
  *
- *  Protected variants should avoid ever throwing an error.
+ *  Protected variants should avoid ever throwing an error.  Must be careful
+ *  to catch errors related to value stack manipulation and property lookup,
+ *  not just the call itself.
+ *
+ *  The only exception is when arguments are insane, e.g. nargs/nrets are out
+ *  of bounds; in such cases an error is thrown for two reasons.  First, we
+ *  can't always respect the value stack input/output guarantees in such cases
+ *  so the caller would end up with the value stack in an unexpected state.
+ *  Second, an attempt to create an error might itself fail (although this
+ *  could be avoided by pushing a preallocated object/string or a primitive
+ *  value).
  */
 
 #include "duk_internal.h"
+
+/*
+ *  Helpers
+ */
+
+struct duk__pcall_prop_args {
+	duk_idx_t obj_idx;
+	duk_idx_t nargs;
+	duk_small_uint_t call_flags;
+};
+typedef struct duk__pcall_prop_args duk__pcall_prop_args;
+
+struct duk__pcall_method_args {
+	duk_idx_t nargs;
+	duk_small_uint_t call_flags;
+};
+typedef struct duk__pcall_method_args duk__pcall_method_args;
+
+struct duk__pcall_args {
+	duk_idx_t nargs;
+	duk_small_uint_t call_flags;
+};
+typedef struct duk__pcall_args duk__pcall_args;
+
+/* Compute and validate idx_func for a certain 'nargs' and 'other'
+ * parameter count (1 or 2, depending on whether 'this' binding is
+ * present).
+ */
+DUK_LOCAL duk_idx_t duk__call_get_idx_func(duk_context *ctx, duk_idx_t nargs, duk_idx_t other) {
+	duk_idx_t idx_func;
+
+	/* XXX: byte arithmetic? */
+
+	DUK_ASSERT(other >= 0);
+
+	idx_func = duk_get_top(ctx) - nargs - other;
+	if (DUK_UNLIKELY((idx_func | nargs) < 0)) {  /* idx_func < 0 || nargs < 0; OR sign bits */
+		DUK_ERROR_TYPE_INVALID_ARGS((duk_hthread *) ctx);
+		/* unreachable */
+	}
+	DUK_ASSERT(duk_is_valid_index(ctx, idx_func));
+	return idx_func;
+}
+
+/* Compute idx_func, assume index will be valid.  This is a valid assumption
+ * for protected calls: nargs < 0 is checked explicitly and duk_safe_call()
+ * validates the argument count.
+ */
+DUK_LOCAL duk_idx_t duk__call_get_idx_func_unvalidated(duk_context *ctx, duk_idx_t nargs, duk_idx_t other) {
+	duk_idx_t idx_func;
+
+	/* XXX: byte arithmetic? */
+
+	DUK_ASSERT(nargs >= 0);
+	DUK_ASSERT(other >= 0);
+
+	idx_func = duk_get_top(ctx) - nargs - other;
+	DUK_ASSERT(idx_func >= 0);
+	DUK_ASSERT(duk_is_valid_index(ctx, idx_func));
+	return idx_func;
+}
 
 /* Prepare value stack for a method call through an object property.
  * May currently throw an error e.g. when getting the property.
  */
 DUK_LOCAL void duk__call_prop_prep_stack(duk_context *ctx, duk_idx_t normalized_obj_idx, duk_idx_t nargs) {
 	DUK_ASSERT_CTX_VALID(ctx);
+	DUK_ASSERT(nargs >= 0);
 
 	DUK_DDD(DUK_DDDPRINT("duk__call_prop_prep_stack, normalized_obj_idx=%ld, nargs=%ld, stacktop=%ld",
 	                     (long) normalized_obj_idx, (long) nargs, (long) duk_get_top(ctx)));
@@ -19,7 +91,7 @@ DUK_LOCAL void duk__call_prop_prep_stack(duk_context *ctx, duk_idx_t normalized_
 
 	/* duplicate key */
 	duk_dup(ctx, -nargs - 1);  /* Note: -nargs alone would fail for nargs == 0, this is OK */
-	duk_get_prop(ctx, normalized_obj_idx);
+	(void) duk_get_prop(ctx, normalized_obj_idx);
 
 	DUK_DDD(DUK_DDDPRINT("func: %!T", (duk_tval *) duk_get_tval(ctx, -1)));
 
@@ -43,23 +115,13 @@ DUK_EXTERNAL void duk_call(duk_context *ctx, duk_idx_t nargs) {
 	DUK_ASSERT_CTX_VALID(ctx);
 	DUK_ASSERT(thr != NULL);
 
-	idx_func = duk_get_top(ctx) - nargs - 1;
-	if (DUK_UNLIKELY(idx_func < 0 || nargs < 0)) {
-		/* note that we can't reliably pop anything here */
-		DUK_ERROR_TYPE_INVALID_ARGS(thr);
-	}
+	idx_func = duk__call_get_idx_func(ctx, nargs, 1);
+	DUK_ASSERT(duk_is_valid_index(ctx, idx_func));
 
-	/* XXX: awkward; we assume there is space for this, overwrite
-	 * directly instead?
-	 */
-	duk_push_undefined(ctx);
-	duk_insert(ctx, idx_func + 1);
+	duk_insert_undefined(ctx, idx_func + 1);
 
 	call_flags = 0;  /* not protected, respect reclimit, not constructor */
-
-	duk_handle_call_unprotected(thr,           /* thread */
-	                            nargs,         /* num_stack_args */
-	                            call_flags);   /* call_flags */
+	duk_handle_call_unprotected(thr, idx_func, call_flags);
 }
 
 DUK_EXTERNAL void duk_call_method(duk_context *ctx, duk_idx_t nargs) {
@@ -70,20 +132,16 @@ DUK_EXTERNAL void duk_call_method(duk_context *ctx, duk_idx_t nargs) {
 	DUK_ASSERT_CTX_VALID(ctx);
 	DUK_ASSERT(thr != NULL);
 
-	idx_func = duk_get_top(ctx) - nargs - 2;  /* must work for nargs <= 0 */
-	if (DUK_UNLIKELY(idx_func < 0 || nargs < 0)) {
-		/* note that we can't reliably pop anything here */
-		DUK_ERROR_TYPE_INVALID_ARGS(thr);
-	}
+	idx_func = duk__call_get_idx_func(ctx, nargs, 2);
+	DUK_ASSERT(duk_is_valid_index(ctx, idx_func));
 
 	call_flags = 0;  /* not protected, respect reclimit, not constructor */
-
-	duk_handle_call_unprotected(thr,           /* thread */
-	                            nargs,         /* num_stack_args */
-	                            call_flags);   /* call_flags */
+	duk_handle_call_unprotected(thr, idx_func, call_flags);
 }
 
 DUK_EXTERNAL void duk_call_prop(duk_context *ctx, duk_idx_t obj_idx, duk_idx_t nargs) {
+	duk_hthread *thr = (duk_hthread *) ctx;
+
 	/*
 	 *  XXX: if duk_handle_call() took values through indices, this could be
 	 *  made much more sensible.  However, duk_handle_call() needs to fudge
@@ -94,107 +152,116 @@ DUK_EXTERNAL void duk_call_prop(duk_context *ctx, duk_idx_t obj_idx, duk_idx_t n
 	DUK_ASSERT_CTX_VALID(ctx);
 
 	obj_idx = duk_require_normalize_index(ctx, obj_idx);  /* make absolute */
+	if (DUK_UNLIKELY(nargs < 0)) {
+		DUK_ERROR_TYPE_INVALID_ARGS(thr);
+	}
 
 	duk__call_prop_prep_stack(ctx, obj_idx, nargs);
 
 	duk_call_method(ctx, nargs);
 }
 
-DUK_EXTERNAL duk_int_t duk_pcall(duk_context *ctx, duk_idx_t nargs) {
+DUK_LOCAL duk_ret_t duk__pcall_raw(duk_context *ctx, void *udata) {
 	duk_hthread *thr = (duk_hthread *) ctx;
-	duk_small_uint_t call_flags;
+	duk__pcall_args *args;
 	duk_idx_t idx_func;
-	duk_int_t rc;
+	duk_int_t ret;
 
 	DUK_ASSERT_CTX_VALID(ctx);
-	DUK_ASSERT(thr != NULL);
+	DUK_ASSERT(udata != NULL);
 
-	idx_func = duk_get_top(ctx) - nargs - 1;  /* must work for nargs <= 0 */
-	if (DUK_UNLIKELY(idx_func < 0 || nargs < 0)) {
-		/* We can't reliably pop anything here because the stack input
-		 * shape is incorrect.  So we throw an error; if the caller has
-		 * no catch point for this, a fatal error will occur.  Another
-		 * alternative would be to just return an error.  But then the
-		 * stack would be in an unknown state which might cause some
-		 * very hard to diagnose problems later on.  Also note that even
-		 * if we did not throw an error here, the underlying call handler
-		 * might STILL throw an out-of-memory error or some other internal
-		 * fatal error.
-		 */
+	args = (duk__pcall_args *) udata;
+	idx_func = duk__call_get_idx_func_unvalidated(ctx, args->nargs, 1);
+	DUK_ASSERT(duk_is_valid_index(ctx, idx_func));
+
+	duk_insert_undefined(ctx, idx_func + 1);
+
+	ret = duk_handle_call_unprotected(thr, idx_func, args->call_flags);
+	DUK_ASSERT(ret == 0);
+	DUK_UNREF(ret);
+
+	return 1;
+}
+
+DUK_EXTERNAL duk_int_t duk_pcall(duk_context *ctx, duk_idx_t nargs) {
+	duk_hthread *thr = (duk_hthread *) ctx;
+	duk__pcall_args args;
+
+	DUK_ASSERT_CTX_VALID(ctx);
+
+	args.nargs = nargs;
+	if (DUK_UNLIKELY(nargs < 0)) {
 		DUK_ERROR_TYPE_INVALID_ARGS(thr);
 		return DUK_EXEC_ERROR;  /* unreachable */
 	}
+	args.call_flags = 0;
 
-	/* Rely on the internal value stack reserve for these operations. */
-	duk_push_undefined(ctx);
-	duk_insert(ctx, idx_func + 1);
+	return duk_safe_call(ctx, duk__pcall_raw, (void *) &args /*udata*/, nargs + 1 /*nargs*/, 1 /*nrets*/);
+}
 
-	call_flags = 0;  /* respect reclimit, not constructor */
+DUK_LOCAL duk_ret_t duk__pcall_method_raw(duk_context *ctx, void *udata) {
+	duk_hthread *thr = (duk_hthread *) ctx;
+	duk__pcall_method_args *args;
+	duk_idx_t idx_func;
+	duk_int_t ret;
 
-	rc = duk_handle_call_protected(thr,           /* thread */
-	                               nargs,         /* num_stack_args */
-	                               call_flags);   /* call_flags */
+	DUK_ASSERT_CTX_VALID(ctx);
+	DUK_ASSERT(udata != NULL);
 
-	return rc;
+	args = (duk__pcall_method_args *) udata;
+
+	idx_func = duk__call_get_idx_func_unvalidated(ctx, args->nargs, 2);
+	DUK_ASSERT(duk_is_valid_index(ctx, idx_func));
+
+	ret = duk_handle_call_unprotected(thr, idx_func, args->call_flags);
+	DUK_ASSERT(ret == 0);
+	DUK_UNREF(ret);
+
+	return 1;
+}
+
+DUK_INTERNAL duk_int_t duk_pcall_method_flags(duk_context *ctx, duk_idx_t nargs, duk_small_uint_t call_flags) {
+	duk_hthread *thr = (duk_hthread *) ctx;
+	duk__pcall_method_args args;
+
+	DUK_ASSERT_CTX_VALID(ctx);
+
+	args.nargs = nargs;
+	if (DUK_UNLIKELY(nargs < 0)) {
+		DUK_ERROR_TYPE_INVALID_ARGS(thr);
+		return DUK_EXEC_ERROR;  /* unreachable */
+	}
+	args.call_flags = call_flags;
+
+	return duk_safe_call(ctx, duk__pcall_method_raw, (void *) &args /*udata*/, nargs + 2 /*nargs*/, 1 /*nrets*/);
 }
 
 DUK_EXTERNAL duk_int_t duk_pcall_method(duk_context *ctx, duk_idx_t nargs) {
-	duk_hthread *thr = (duk_hthread *) ctx;
-	duk_small_uint_t call_flags;
-	duk_idx_t idx_func;
-	duk_int_t rc;
-
-	DUK_ASSERT_CTX_VALID(ctx);
-	DUK_ASSERT(thr != NULL);
-
-	idx_func = duk_get_top(ctx) - nargs - 2;  /* must work for nargs <= 0 */
-	if (DUK_UNLIKELY(idx_func < 0 || nargs < 0)) {
-		/* See comments in duk_pcall(). */
-		DUK_ERROR_TYPE_INVALID_ARGS(thr);
-		return DUK_EXEC_ERROR;  /* unreachable */
-	}
-
-	call_flags = 0;  /* respect reclimit, not constructor */
-
-	rc = duk_handle_call_protected(thr,           /* thread */
-	                               nargs,         /* num_stack_args */
-	                               call_flags);   /* call_flags */
-
-	return rc;
+	return duk_pcall_method_flags(ctx, nargs, 0);
 }
 
-struct duk__pcall_prop_args {
-	duk_idx_t obj_idx;
-	duk_idx_t nargs;
-};
-typedef struct duk__pcall_prop_args duk__pcall_prop_args;
-
 DUK_LOCAL duk_ret_t duk__pcall_prop_raw(duk_context *ctx, void *udata) {
-	duk_idx_t obj_idx;
-	duk_idx_t nargs;
 	duk__pcall_prop_args *args;
+	duk_idx_t obj_idx;
+	duk_int_t ret;
 
 	DUK_ASSERT_CTX_VALID(ctx);
 	DUK_ASSERT(udata != NULL);
 
 	args = (duk__pcall_prop_args *) udata;
-	obj_idx = args->obj_idx;
-	nargs = args->nargs;
 
-	obj_idx = duk_require_normalize_index(ctx, obj_idx);  /* make absolute */
-	duk__call_prop_prep_stack(ctx, obj_idx, nargs);
-	duk_call_method(ctx, nargs);
+	obj_idx = duk_require_normalize_index(ctx, args->obj_idx);  /* make absolute */
+	duk__call_prop_prep_stack(ctx, obj_idx, args->nargs);
+
+	ret = duk_handle_call_unprotected_nargs((duk_hthread *) ctx, args->nargs, args->call_flags);
+	DUK_ASSERT(ret == 0);
+	DUK_UNREF(ret);
 	return 1;
 }
 
 DUK_EXTERNAL duk_int_t duk_pcall_prop(duk_context *ctx, duk_idx_t obj_idx, duk_idx_t nargs) {
 	duk_hthread *thr = (duk_hthread *) ctx;
 	duk__pcall_prop_args args;
-
-	/*
-	 *  Must be careful to catch errors related to value stack manipulation
-	 *  and property lookup, not just the call itself.
-	 */
 
 	DUK_ASSERT_CTX_VALID(ctx);
 
@@ -204,11 +271,8 @@ DUK_EXTERNAL duk_int_t duk_pcall_prop(duk_context *ctx, duk_idx_t obj_idx, duk_i
 		DUK_ERROR_TYPE_INVALID_ARGS(thr);
 		return DUK_EXEC_ERROR;  /* unreachable */
 	}
+	args.call_flags = 0;
 
-	/* Inputs: explicit arguments (nargs), +1 for key.  If the value stack
-	 * does not contain enough args, an error is thrown; this matches
-	 * behavior of the other protected call API functions.
-	 */
 	return duk_safe_call(ctx, duk__pcall_prop_raw, (void *) &args /*udata*/, nargs + 1 /*nargs*/, 1 /*nrets*/);
 }
 
@@ -219,8 +283,17 @@ DUK_EXTERNAL duk_int_t duk_safe_call(duk_context *ctx, duk_safe_call_function fu
 	DUK_ASSERT_CTX_VALID(ctx);
 	DUK_ASSERT(thr != NULL);
 
-	if (DUK_UNLIKELY(duk_get_top(ctx) < nargs || nargs < 0 || nrets < 0)) {
-		/* See comments in duk_pcall(). */
+	/* nargs condition; fail if: top - bottom < nargs
+	 *                      <=>  top < bottom + nargs
+	 * nrets condition; fail if: end - (top - nargs) < nrets
+	 *                      <=>  end - top + nargs < nrets
+	 *                      <=>  end + nargs < top + nrets
+	 */
+	/* XXX: check for any reserve? */
+
+	if (DUK_UNLIKELY((nargs | nrets) < 0 ||  /* nargs < 0 || nrets < 0; OR sign bits */
+	                 thr->valstack_top < thr->valstack_bottom + nargs ||        /* nargs too large compared to top */
+	                 thr->valstack_end + nargs < thr->valstack_top + nrets)) {  /* nrets too large compared to reserve */
 		DUK_ERROR_TYPE_INVALID_ARGS(thr);
 		return DUK_EXEC_ERROR;  /* unreachable */
 	}
@@ -240,16 +313,13 @@ DUK_EXTERNAL void duk_new(duk_context *ctx, duk_idx_t nargs) {
 
 	DUK_ASSERT_CTX_VALID(ctx);
 
-	idx_func = duk_get_top(ctx) - nargs - 1;
-	if (DUK_UNLIKELY(idx_func < 0 || nargs < 0)) {
-		/* note that we can't reliably pop anything here */
-		DUK_ERROR_TYPE_INVALID_ARGS(thr);
-	}
+	idx_func = duk__call_get_idx_func(ctx, nargs, 1);
+	DUK_ASSERT(duk_is_valid_index(ctx, idx_func));
 
 	duk_push_object(ctx);  /* default instance; internal proto updated by call handling */
 	duk_insert(ctx, idx_func + 1);
 
-	duk_handle_call_unprotected((duk_hthread *) ctx, nargs, DUK_CALL_FLAG_CONSTRUCTOR_CALL);
+	duk_handle_call_unprotected(thr, idx_func, DUK_CALL_FLAG_CONSTRUCTOR_CALL);
 }
 
 DUK_LOCAL duk_ret_t duk__pnew_helper(duk_context *ctx, void *udata) {