+++ tests/perf/test-empty-loop-step3.js
@@ -1,0 +1,15 @@
+if (typeof print !== 'function') { print = console.log; }
+
+function test() {
+    var i;
+
+    for (i = 0; i < 3e8; i += 3) {
+    }
+}
+
+try {
+    test();
+} catch (e) {
+    print(e.stack || e);
+    throw e;
+}